requirements.txt
beautifulsoup4==4.12.3
chardet==5.2.0
click==8.1.8
grpcio==1.69.0
grpcio-tools==1.69.0
importlib_metadata==6.8.0
lxml==5.1.0
numpy==1.24.4
opencv-python==4.6.0.66
packaging==23.2
pillow==10.2.0
plum-dispatch==1.7.4
protobuf==5.29.3
pybit7z==0.4.0
PyMuPDF>=1.26.0,<1.27.0
pytesseract==0.3.10
python-docx==1.1.0
python-magic==0.4.27
snowflake-id==1.0.0
soupsieve==2.5
typing_extensions>=4.10.0
zipp==3.17.0
zxing-cpp==2.1.0
msoffcrypto-tool==5.4.1
olefile==0.47
easyocr==1.7.0
loguru==0.7.3
pytest==8.3.4
pytest-cov==6.0.0